+++ test/options.test.js
@@ -21,48 +21,6 @@ test('options: `includeResponseBody`, custom `headers`, `descriptionLength`, `en
   }
 })
 
-// test('option: `ensureSecureImageRequest` edge cases', async () => {
-//   const url = 'http://news.bbc.co.uk '
-//   try {
-//     const metadata = await urlMetadata(url, {
-//       requestHeaders: {
-//         'User-Agent': 'sparky',
-//         From: 'foobarred@foo.com'
-//       },
-//       ensureSecureImageRequest: true
-//     })
-//     // this page has favicons with `//:` protocol-relative cases
-//     // with no scheme; ensure they get upgraded to 'https://'
-//     metadata.favicons.forEach((favicon) => {
-//       expect(favicon.href.indexOf('https://')).toBe(0)
-//     })
-//     // this page also has img tags with base64 data URIs
-//     // make sure those are untouched but other imgs are `https://`
-//     metadata.imgTags.forEach((img) => {
-//       const httpsIndex = img.src.indexOf('https://')
-//       const dataUriIndex = img.src.indexOf('data:')
-//       expect(httpsIndex === 0 || dataUriIndex === 0).toBe(true)
-//     })
-//   } catch (err) {
-//     expect(err).toBe(undefined)
-//   }
-// })
-
-test('option: `parseResponseObject`', async () => {
-  try {
-    const url = 'https://www.npmjs.com/package/url-metadata'
-    const response = await fetch(url)
-    // pass null `url` param & response object as option`
-    const metadata = await urlMetadata(null, { parseResponseObject: response })
-    expect(metadata.url).toBe(url)
-    expect(metadata.title).toBe('url-metadata - npm')
-    expect(metadata.lang).toBe('en')
-    expect(metadata.charset).toBe('utf-8')
-  } catch (e) {
-    expect(e).toBe(undefined)
-  }
-})
-
 test('option: `parseResponseObject` from html string', async () => {
   const html = `
   <!DOCTYPE html>
@@ -94,5 +52,17 @@ test('option: `parseResponseObject` from html string', async () => {
     expect(metadata.author).toBe('foobar')
   } catch (e) {
     expect(e).toBe(undefined)
+  }
+})
+
+const size = 1000
+test('option: max `size` 1000 bytes aborts call & errors', async () => {
+  try {
+    const url = 'https://google.com'
+    const metadata = await urlMetadata(url, { size })
+    // should not reach here, but just in case:
+    expect(metadata).toBeUndefined()
+  } catch (e) {
+    expect(e.message).toContain(`over limit: ${size}`)
   }
 })